EDIT FOR TASK 1

+++ alx_book_store.sql
@@ -39,7 +39,7 @@ CREATE TABLE IF NOT EXISTS Order_Details (
     orderdetail_id INT PRIMARY KEY AUTO_INCREMENT,
     order_id INT NOT NULL,
     book_id INT NOT NULL,
-    quantity INT NOT NULL,
+    quantity DOUBLE NOT NULL,
     FOREIGN KEY (order_id) REFERENCES Orders(order_id),
     FOREIGN KEY (book_id) REFERENCES Books(book_id)
 );
